merge_streams/requirements.txt
pipelinewise-singer-python==1.2.0
